+++ spring/loja/src/main/java/com/br/maskotes/loja/service/FornecedorServiceImpl.java
@@ -6,7 +6,6 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
-
 import com.br.maskotes.loja.entitites.Fornecedor;
 import com.br.maskotes.loja.repository.FornecedorRepository;
 
@@ -23,10 +22,15 @@ public class FornecedorServiceImpl implements FornecedorService {
         return fornecedorRepository.findAll();
     }
 
-    @Override
-    public Fornecedor listOne(@PathVariable Long id) {
+   @Override
+public Fornecedor listOne(@PathVariable Long id) {
+    if (id != null) {
         return fornecedorRepository.findById(id).orElse(null);
+    } else {
+        throw new IllegalArgumentException("O ID não pode ser nulo.");
     }
+}
+
 
     @Override
     public Fornecedor create(Fornecedor fornecedor) {
@@ -45,7 +49,16 @@ public class FornecedorServiceImpl implements FornecedorService {
     }
 
     @Override
-    public void delete(Long id) {
+public void delete(Long id) {
+    if (id != null) {
         fornecedorRepository.deleteById(id);
+    } else {
+        throw new IllegalArgumentException("O ID não pode ser nulo.");
+    }
+}
+
+       @Override
+    public List<Fornecedor> findByNomeContainingIgnoreCase(String name) {
+        return fornecedorRepository.findByNomeContainingIgnoreCase(name);
     }
 }
