+++ spring/loja/src/main/java/com/br/maskotes/loja/service/ProdutoVendidoServiceImpl.java
@@ -1,10 +1,12 @@
 package com.br.maskotes.loja.service;
 
+
 import com.br.maskotes.loja.entitites.ProdutoVendido;
 import com.br.maskotes.loja.repository.ProdutoVendidoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
 
@@ -19,15 +21,25 @@ public class ProdutoVendidoServiceImpl implements ProdutoVendidoService {
         return produtoVendidoRepository.findAll();
     }
 
-    @Override
-    public ProdutoVendido listOne(Long id) {
+ @Override
+public ProdutoVendido listOne(@PathVariable Long id) {
+    if (id != null) {
         return produtoVendidoRepository.findById(id).orElse(null);
+    } else {
+        throw new IllegalArgumentException("O ID não pode ser nulo.");
     }
+}
 
-    @Override
-    public ProdutoVendido create(ProdutoVendido produtoVendido) {
+
+@Override
+public ProdutoVendido create(ProdutoVendido produtoVendido) {
+    if (produtoVendido != null) {
         return produtoVendidoRepository.save(produtoVendido);
+    } else {
+        throw new IllegalArgumentException("O produtoVendido não pode ser nulo.");
     }
+}
+
 
     @Override
     public ProdutoVendido update(ProdutoVendido produtoVendido) {
@@ -39,6 +51,11 @@ public class ProdutoVendidoServiceImpl implements ProdutoVendidoService {
 
     @Override
     public void delete(Long id) {
-        produtoVendidoRepository.deleteById(id);
+        if (id != null) {
+            produtoVendidoRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("O ID não pode ser nulo.");
+        }
     }
+    
 }
