+++ spring/loja/src/main/java/com/br/maskotes/loja/service/TipoProdutoServiceImpl.java
@@ -5,6 +5,7 @@ import com.br.maskotes.loja.repository.TipoProdutoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.List;
 
@@ -20,10 +21,15 @@ public class TipoProdutoServiceImpl implements TipoProdutoService {
         return tipoProdutoRepository.findAll();
     }
 
-    @Override
-    public TipoProduto listOne(Long id) {
+   @Override
+public TipoProduto listOne(@PathVariable Long id) {
+    if (id != null) {
         return tipoProdutoRepository.findById(id).orElse(null);
+    } else {
+        throw new IllegalArgumentException("O ID não pode ser nulo.");
     }
+}
+
 
     @Override
     public TipoProduto create(TipoProduto tipoProduto) {
@@ -43,6 +49,16 @@ public class TipoProdutoServiceImpl implements TipoProdutoService {
 
     @Override
     public void delete(Long id) {
-        tipoProdutoRepository.deleteById(id);
+        if (id != null) {
+            tipoProdutoRepository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("O ID não pode ser nulo.");
+        }
+    }
+    
+
+    @Override
+    public List<TipoProduto> findByDescricaoContainingIgnoreCase(String nome) {
+        return tipoProdutoRepository.findByDescricaoContainingIgnoreCase(nome);
     }
 }
