+++ spring/loja/src/main/java/com/br/maskotes/loja/service/VendaServiceImpl.java
@@ -1,11 +1,12 @@
 package com.br.maskotes.loja.service;
 
+import java.util.Date;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
-
+import org.springframework.web.bind.annotation.RequestParam;
 import com.br.maskotes.loja.entitites.Venda;
 import com.br.maskotes.loja.repository.VendaRepository;
 
@@ -22,10 +23,15 @@ public class VendaServiceImpl implements VendaService {
         return vendaRepository.findAll();
     }
 
-    @Override
-    public Venda listOne(@PathVariable Long id) {
+   @Override
+public Venda listOne(@PathVariable Long id) {
+    if (id != null) {
         return vendaRepository.findById(id).orElse(null);
+    } else {
+        throw new IllegalArgumentException("O ID não pode ser nulo.");
     }
+}
+
 
     @Override
     public Venda create(Venda venda) {
@@ -44,7 +50,22 @@ public class VendaServiceImpl implements VendaService {
     }
 
     @Override
-    public void delete(Long id) {
+public void delete(Long id) {
+    if (id != null) {
         vendaRepository.deleteById(id);
+    } else {
+        throw new IllegalArgumentException("O ID não pode ser nulo.");
+    }
+}
+
+
+    @Override
+    public List<Venda> findAllByDataVenda(@PathVariable Date data) {
+        return vendaRepository.findAllBydataVenda(data);
+    }
+
+    @Override
+    public List<Venda> findAllByDataVendaBetween(@RequestParam Date inicio,@RequestParam Date fim) {
+        return vendaRepository.findAllByDataVendaBetween(inicio, fim);
     }
 }
